requirements.txt
requests
bs4
flask
groq
numpy
python-dotenv
gunicorn